temp 반영 커밋

+++ hellong/src/main/webapp/WEB-INF/views/SH-Views/QuestionContent.jsp
@@ -1,6 +1,7 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8"
 	pageEncoding="UTF-8"%>
 <%@ include file="../header.jsp"%>
+
 <!DOCTYPE html>
 <html lang="ko">
 <head>
@@ -25,13 +26,18 @@
 							var frmData = $('#comment_form').serialize();
 							// jQuery의 $.post 메소드를 사용해 서버로 데이터 전송
 		$.post("commentInsert",frmData,function(board){
-							var newCommentHtml = '<li>댓글 : '+ board.b_content
-																		+ ' <a href="/deleteComment?Comm_number='
-																		+ board.b_number
-																		+ '&bId='
-																		+ board.b_comm_group
-																		+ '" onclick="return confirm(\'댓글을 삭제하시겠습니까?\');">삭제</a></li>';
-																$('.list_comments').prepend(newCommentHtml); // 새로운 <li>를 list_comments ul에 추가
+							var newCommentHtml ='<li>' +
+						    '<div class="comment-details">' +
+						    '<h4 class="comment-author" style="font-size: 11px">' +
+						    (board.m_number != board.b_comm_group ?  '관리자' : '질문자') +
+						    '</h4>' +
+						    '<p class="comment-description" style="font-size: 18px;">' + board.b_content + '</p>' +
+						    '<span style="font-size: 12px; color: #757575;">' + board.b_regdate + '</span>' +
+						    '<div>' +
+						    '<a href="/deleteComment?Comm_number=' + board.b_number + '&bId=' + board.b_comm_group + '" onclick="return confirm(\'댓글을 삭제하시겠습니까?\');">삭제</a>' +
+						    '</div>' +
+						    '</div>' +
+						    '</li><hr>';																$('.list_comments').prepend(newCommentHtml); // 새로운 <li>를 list_comments ul에 추가
 																$('[name="comment_body"]').val(''); // 댓글 입력란을 비웁니다.
 															}).fail(
 															function() {alert('댓글 등록에 실패했습니다.');});
@@ -145,4 +151,5 @@
 	<script type="text/javascript"
 		src="<%=request.getContextPath()%>/js/index.js"></script>
 </body>
+<%@ include file="../footer.jsp"%>
 </html>
